+++ src/hm/hmSystem.h
@@ -16,8 +16,8 @@ class HmSystem {
         HmSystem() {}
 
         void setup(uint32_t *timestamp, cfgInst_t *config, IApp *app) {
-            INVERTERTYPE::timestamp = timestamp;
-            INVERTERTYPE::generalConfig = config;
+            INVERTERTYPE::Timestamp = timestamp;
+            INVERTERTYPE::GeneralConfig = config;
             INVERTERTYPE::App = app;
             //mInverter[0].app           = app;
         }
@@ -26,7 +26,7 @@ class HmSystem {
             DPRINTLN(DBG_VERBOSE, F("hmSystem.h:addInverter"));
             INVERTERTYPE *iv = &mInverter[id];
             iv->id         = id;
-            iv->config     = &mInverter[0].generalConfig->iv[id];
+            iv->config     = &INVERTERTYPE::GeneralConfig->iv[id];
             DPRINT(DBG_VERBOSE, "SERIAL: " + String(iv->config->serial.b[5], HEX));
             DPRINTLN(DBG_VERBOSE, " " + String(iv->config->serial.b[4], HEX));
             if((iv->config->serial.b[5] == 0x11) || (iv->config->serial.b[5] == 0x10)) {