+++ src/plugins/history.h
@@ -57,17 +57,15 @@ class HistoryData {
 
         void tickerSecond() {
             float curPwr = 0;
-            //float maxPwr = 0;
             float yldDay = -0.1;
             uint32_t ts = 0;
 
             for (uint8_t i = 0; i < mSys->getNumInverters(); i++) {
                 Inverter<> *iv = mSys->getInverterByPos(i);
-                record_t<> *rec = iv->getRecordStruct(RealTimeRunData_Debug);
                 if (iv == NULL)
                     continue;
+                record_t<> *rec = iv->getRecordStruct(RealTimeRunData_Debug);
                 curPwr += iv->getChannelFieldValue(CH0, FLD_PAC, rec);
-                //maxPwr += iv->getChannelFieldValue(CH0, FLD_MP, rec);
                 yldDay += iv->getChannelFieldValue(CH0, FLD_YD, rec);
                 if (rec->ts > ts)
                     ts = rec->ts;
@@ -81,8 +79,6 @@ class HistoryData {
                     if (curPwr > mMaximumDay)
                         mMaximumDay = roundf(curPwr);
                 }
-                //if (maxPwr > 0)
-                //    mMaximumDay = roundf(maxPwr);
             }
 
             if ((++mCurPwrDay.loopCnt % mCurPwrDay.refreshCycle) == 0) {