+++ src/publisher/pubMqttIvData.h
@@ -24,7 +24,8 @@ class PubMqttIvData {
     public:
         PubMqttIvData() : mTotal{}, mSubTopic{}, mVal{} {}
 
-        void setup(HMSYSTEM *sys, bool json, uint32_t *utcTs, std::queue<sendListCmdIv> *sendList) {
+        void setup(IApp *app, HMSYSTEM *sys, uint32_t *utcTs, std::queue<sendListCmdIv> *sendList) {
+            mApp           = app;
             mSys           = sys;
             mJson          = json;
             mUtcTimestamp  = utcTs;
@@ -169,9 +170,6 @@ class PubMqttIvData {
                                 case FLD_PDC:
                                     mTotal[3] += mIv->getValue(mPos, rec);
                                     break;
-                                case FLD_MP:
-                                    mTotal[4] += mIv->getValue(mPos, rec);
-                                    break;
                             }
                         } else
                             mAllTotalFound = false;
@@ -299,6 +297,7 @@ class PubMqttIvData {
                     case 4:
                         fieldId = FLD_MP;
                         retained = false;
+                        mTotal[4] = mApp->getTotalMaxPower();
                         break;
                 }
                 if (!mJson) {
@@ -327,6 +326,9 @@ class PubMqttIvData {
                 mState = IDLE;
             }
         }
+
+    private:
+        IApp *mApp = nullptr;
 
         HMSYSTEM *mSys = nullptr;
         uint32_t *mUtcTimestamp = nullptr;