+++ src/web/web.h
@@ -498,12 +498,13 @@ class Web {
 
             if (request->arg("invInterval") != "")
                 mConfig->inst.sendInterval = request->arg("invInterval").toInt();
-            mConfig->inst.rstYieldMidNight = (request->arg("invRstMid") == "on");
+            mConfig->inst.rstValsAtMidNight = (request->arg("invRstMid") == "on");
             mConfig->inst.rstValsCommStop = (request->arg("invRstComStop") == "on");
+            mConfig->inst.rstValsCommStart = (request->arg("invRstComStart") == "on");
             mConfig->inst.rstValsNotAvail = (request->arg("invRstNotAvail") == "on");
             mConfig->inst.startWithoutTime = (request->arg("strtWthtTm") == "on");
             mConfig->inst.readGrid = (request->arg("rdGrid") == "on");
-            mConfig->inst.rstMaxValsMidNight = (request->arg("invRstMaxMid") == "on");
+            mConfig->inst.rstIncludeMaxVals = (request->arg("invRstMaxMid") == "on");
 
 
             // pinout